+++ test/pg_test.rb
@@ -5,7 +5,7 @@ class PgTest < Minitest::Test
     conn.exec("DELETE FROM pg_items")
   end
 
-  def test_text
+  def test_vector_text
     embedding = [1.5, 2, 3]
     conn.exec_params("INSERT INTO pg_items (embedding) VALUES ($1), (NULL)", [embedding])
 
@@ -14,13 +14,40 @@ class PgTest < Minitest::Test
     assert_nil res[1]["embedding"]
   end
 
-  def test_binary
+  def test_vector_binary
     embedding = [1.5, 2, 3]
     conn.exec_params("INSERT INTO pg_items (embedding) VALUES ($1), (NULL)", [embedding])
 
     res = conn.exec_params("SELECT * FROM pg_items ORDER BY id", [], 1).to_a
     assert_equal embedding, res[0]["embedding"]
     assert_nil res[1]["embedding"]
+  end
+
+  def test_halfvec_text
+    embedding = [1.5, 2, 3]
+    conn.exec_params("INSERT INTO pg_items (half_embedding) VALUES ($1), (NULL)", [embedding])
+
+    res = conn.exec("SELECT * FROM pg_items ORDER BY id").to_a
+    assert_equal embedding, res[0]["half_embedding"]
+    assert_nil res[1]["half_embedding"]
+  end
+
+  def test_sparsevec_text
+    embedding = Pgvector::SparseVector.from_dense([1.5, 2, 3])
+    conn.exec_params("INSERT INTO pg_items (sparse_embedding) VALUES ($1), (NULL)", [embedding])
+
+    res = conn.exec("SELECT * FROM pg_items ORDER BY id").to_a
+    assert_equal [1.5, 2, 3], res[0]["sparse_embedding"].to_a
+    assert_nil res[1]["sparse_embedding"]
+  end
+
+  def test_sparsevec_binary
+    embedding = Pgvector::SparseVector.from_dense([1.5, 2, 3])
+    conn.exec_params("INSERT INTO pg_items (sparse_embedding) VALUES ($1), (NULL)", [embedding])
+
+    res = conn.exec_params("SELECT * FROM pg_items ORDER BY id", [], 1).to_a
+    assert_equal [1.5, 2, 3], res[0]["sparse_embedding"].to_a
+    assert_nil res[1]["sparse_embedding"]
   end
 
   def conn
@@ -31,7 +58,7 @@ class PgTest < Minitest::Test
         conn.exec("CREATE EXTENSION IF NOT EXISTS vector")
       end
       conn.exec("DROP TABLE IF EXISTS pg_items")
-      conn.exec("CREATE TABLE pg_items (id bigserial PRIMARY KEY, embedding vector(3))")
+      conn.exec("CREATE TABLE pg_items (id bigserial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3), sparse_embedding sparsevec(3))")
 
       registry = PG::BasicTypeRegistry.new.define_default_types
       Pgvector::PG.register_vector(registry)