+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.3 (unreleased)
 
+- Added support for `halfvec` and `sparsevec` types
 - Added `taxicab` distance for Sequel
 
 ## 0.2.2 (2023-10-03)